Change 'edit' functionality around todo name
When a user clicks on an existing todo, the name of the todo will not
editable on default. The user can click the 'edit' button next to the name
of the todo if they wish to change the todo.

This will be duplicated with the todo description in a future commit.