Fix bug re: deleting project w/ todos
Previously, if a project that had todos under it was deleted, thus moving
those todos under the 'Personal' project, the todos were not being
re-stored in localStorage. This led to the todos still being listed under
the now-deleted project in localStorage, which led to them not showing up
under any Project tab when the page is reloaded. This commit corrects that